Move memoize to its own file
